+++ packages/retail-ui/components/Fias/Form/FiasForm.tsx
@@ -4,7 +4,7 @@ import Button from '../../Button';
 import { locale } from '../../LocaleProvider/decorators';
 import { FiasLocale, FiasLocaleHelper } from '../locale';
 import { FiasComboBox, FiasComboBoxChangeEvent, FiasComboBoxProps } from './FiasComboBox';
-import styles from './FiasForm.less';
+import styles from './FiasForm.module.less';
 import {
   Fields,
   FormValidation,
@@ -324,17 +324,18 @@ export class FiasForm extends React.Component<FiasFormProps, FiasFormState> {
 
     const renderItem = (address: Address): string => {
       const element = address.fields[field];
-      const hasParents = Boolean(address.getClosestParentFiasId(field));
-
-      const fieldText = element ? element.getText(!hasParents && element.isTypeMatchField(field)) : '';
-
-      if (field === Fields.region && element && element.data) {
-        const regionCode = element.data.code.substr(0, 2);
-        return `${regionCode} ${fieldText}`;
+      if (!element) {
+        return '';
       }
-      // TODO: handle possible identical texts of elements
-      // while in the "not directParent" search mode
-      return hasParents ? [address.getText(field), fieldText].join(', ') : fieldText;
+
+      if (field === Fields.region && element.data) {
+        const regionCode = element.data.code.substr(0, 2);
+        return `${regionCode} ${element.getText()}`;
+      }
+
+      const diffAddress = new Address({ fields: this.state.address.getDiffFields(address, this.props.fieldsSettings) });
+      const hasParentFields = Boolean(diffAddress.getClosestParentFiasId(field));
+      return hasParentFields ? diffAddress.getText() : element.getText(element.isTypeMatchField(field));
     };
 
     const renderValue = (address: Address): React.ReactNode => {
@@ -402,7 +403,7 @@ export class FiasForm extends React.Component<FiasFormProps, FiasFormState> {
     };
   }
 
-  private createItemsSource = async (searchText: string, field?: Fields) => {
+  private createItemsSource = async (searchText: string, field: Fields) => {
     const { address } = this.state;
     const limit = this.props.limit || FiasForm.defaultProps.limit;
 
@@ -412,7 +413,7 @@ export class FiasForm extends React.Component<FiasFormProps, FiasFormState> {
         field,
         parentFiasId: address.getClosestParentFiasId(field),
         fullAddress: address.isAllowedToSearchFullAddress(field),
-        directParent: !address.isAllowedToSearchThroughChildrenOfDirectParent(field),
+        directParent: !address.isAllowedToSearchThroughChildrenOfDirectParent(field, this.props.fieldsSettings),
         limit: limit + 1, // +1 to detect if there are more items
       };
       return this.props.api.search(options).then(result => {