+++ packages/retail-ui/components/Modal/ModalFooter.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import getScrollWidth from '../../lib/dom/getScrollWidth';
 import Sticky from '../Sticky/Sticky';
 import { ModalContext } from './ModalContext';
-import styles from './Modal.less';
+import styles from './Modal.module.less';
 import { cx } from '../../lib/theming/Emotion';
 import jsStyles from './Modal.styles';
 import { ThemeConsumer } from '../internal/ThemeContext';