+++ packages/retail-ui/components/Paging/Paging.styles.ts
@@ -1,5 +1,5 @@
 import { css } from '../../lib/theming/Emotion';
-import styles from './Paging.less';
+import styles from './Paging.module.less';
 import { ITheme } from '../../lib/theming/Theme';
 
 const jsStyles = {