+++ packages/retail-ui/components/Tabs/Tabs.tsx
@@ -4,7 +4,7 @@ import * as PropTypes from 'prop-types';
 import Indicator from './Indicator';
 import { TabsContext } from './TabsContext';
 import { TabProps, TabWithContext, Tab } from './Tab';
-import styles from './Tabs.less';
+import styles from './Tabs.module.less';
 import { cx } from '../../lib/theming/Emotion';
 
 export interface TabsProps {