+++ packages/retail-ui/components/Toast/ToastView.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import * as PropTypes from 'prop-types';
 import CROSS from '../internal/cross';
 import ZIndex from '../ZIndex/ZIndex';
-import styles from './ToastView.less';
+import styles from './ToastView.module.less';
 import { Nullable } from '../../typings/utility-types';
 import jsStyles from './ToastView.styles';
 import { cx } from '../../lib/theming/Emotion';