+++ packages/retail-ui/components/TokenInput/TokenInput.styles.ts
@@ -1,5 +1,5 @@
 import { css } from '../../lib/theming/Emotion';
-import styles from './TokenInput.less';
+import styles from './TokenInput.module.less';
 import { ITheme } from '../../lib/theming/Theme';
 
 const jsStyles = {